+++ sus/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: otchekai <otchekai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 15:39:10 by otchekai          #+#    #+#             */
-/*   Updated: 2022/10/24 07:28:48 by otchekai         ###   ########.fr       */
+/*   Updated: 2022/11/03 12:27:23 by otchekai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
